+++ apps/ledger-live-desktop/src/renderer/families/solana/SendAmountFields.tsx
@@ -1,9 +1,15 @@
-import React from "react";
+import React, { useMemo } from "react";
 import Alert from "~/renderer/components/Alert";
 import { Flex } from "@ledgerhq/react-ui";
 import TranslatedError from "~/renderer/components/TranslatedError";
 import { Account } from "@ledgerhq/types-live";
-import type { Transaction, TransactionStatus } from "@ledgerhq/live-common/families/solana/types";
+import type {
+  Transaction,
+  TransactionStatus,
+  SolanaTokenAccount,
+} from "@ledgerhq/live-common/families/solana/types";
+import { findSubAccountById } from "@ledgerhq/coin-framework/account/helpers";
+import TokenTransferFeesWarning from "./Token2022/TokenTransferFeesWarning";
 
 type Props = {
   account: Account;
@@ -20,13 +26,24 @@ type Props = {
  * to complete a transaction.
  *
  */
-const SendAmountFields = ({ status }: Props) => {
-  if (!status.errors.fee) return null;
+const SendAmountFields = ({ account, transaction, status }: Props) => {
+  const tokenAccount = useMemo(
+    () =>
+      transaction.subAccountId
+        ? (findSubAccountById(account, transaction.subAccountId) as SolanaTokenAccount)
+        : undefined,
+    [account, transaction.subAccountId],
+  );
   return (
     <Flex>
-      <Alert type="warning" data-testid="insufficient-funds-warning">
-        <TranslatedError error={status.errors.fee} />
-      </Alert>
+      {status.errors.fee ? (
+        <Alert type="warning" data-testid="insufficient-funds-warning">
+          <TranslatedError error={status.errors.fee} />
+        </Alert>
+      ) : null}
+      {tokenAccount ? (
+        <TokenTransferFeesWarning tokenAccount={tokenAccount} transaction={transaction} />
+      ) : null}
     </Flex>
   );
 };