+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -129,6 +129,8 @@ export type CurrencyFeatures = {
   currencyMantra: DefaultFeature;
   currencyXion: DefaultFeature;
   currencyZenrock: DefaultFeature;
+  currencySonic: DefaultFeature;
+  currencySonicBlaze: DefaultFeature;
 };
 
 /**