+++ libs/domain-service/CHANGELOG.md
@@ -1,11 +1,11 @@
 # @ledgerhq/domain-service
 
-## 1.2.20-nightly.0
+## 1.2.20-next.0
 
 ### Patch Changes
 
-- Updated dependencies [[`ede6db0`](https://github.com/LedgerHQ/ledger-live/commit/ede6db0b94193cc9072aeb87e90f4098f0434af0)]:
-  - @ledgerhq/types-live@6.61.0-nightly.0
+- Updated dependencies [[`e4d9a5c`](https://github.com/LedgerHQ/ledger-live/commit/e4d9a5ce6c3e8f2b3829f8f5772e7ba712a4a50c)]:
+  - @ledgerhq/types-live@6.61.0-next.0
 
 ## 1.2.19